improve musicview's inheritance from abstractviewpanel
The abstract actions class now uses the inherited Actions class from
the subclasses. So far they are practically empty and only call
super() with the Actions class from the viewers module. But this will
change once we discern between generic and specific actions.
